Move all logic to server class
