+++ core/src/Http/Server.h
@@ -1,0 +1,33 @@
+#pragma once
+
+#include <memory>
+#include <string>
+#include <types.h>
+
+#include <Http/Request.h>
+#include <Http/Response.h>
+
+namespace http_get { inline namespace Http {
+
+class Server {
+public:
+    using handler_t = std::function<Response(Request *)>;
+
+    virtual bool start() noexcept = 0;
+    virtual Server *addHandler(const std::string &pattern,
+                               handler_t &&handler) noexcept = 0;
+
+    // constructors
+public:
+    struct Settings {
+        std::string address = "0.0.0.0";
+        u16 port = 8080;
+    };
+
+    static std::unique_ptr<Server> create(const Settings &settings,
+                                          std::atomic_flag &m_stop);
+
+    virtual ~Server() {}
+};
+
+}} // namespace http_get::Http
